extract first message in standalone file
